fix: duplicate open in vscode command registration
