+++ src/vs/sessions/contrib/chat/browser/openInVSCode.contribution.ts
@@ -12,7 +12,7 @@ import { Action2, registerAction2 } from '../../../../platform/actions/common/ac
 import { AGENT_HOST_SCHEME, fromAgentHostUri } from '../../../../platform/agentHost/common/agentHostUri.js';
 import { IRemoteAgentHostService } from '../../../../platform/agentHost/common/remoteAgentHostService.js';
 import { ContextKeyExpr } from '../../../../platform/contextkey/common/contextkey.js';
-import { INativeHostService } from '../../../../platform/native/common/native.js';
+import { IOpenerService } from '../../../../platform/opener/common/opener.js';
 import { IProductService } from '../../../../platform/product/common/productService.js';
 import { ITelemetryService } from '../../../../platform/telemetry/common/telemetry.js';
 import { IsAuxiliaryWindowContext } from '../../../../workbench/common/contextkeys.js';
@@ -22,14 +22,12 @@ import { Menus } from '../../../browser/menus.js';
 import { CopilotCLISessionType } from '../../../services/sessions/common/session.js';
 import { ISessionsManagementService } from '../../../services/sessions/common/sessionsManagement.js';
 import { ISessionsProvidersService } from '../../../services/sessions/browser/sessionsProvidersService.js';
-import { resolveRemoteAuthority } from '../browser/openInVSCodeUtils.js';
+import { resolveRemoteAuthority } from './openInVSCodeUtils.js';
 
 /**
- * Desktop version of the "Open in VS Code" action.
- *
- * Launches the host VS Code app via {@link INativeHostService.launchSiblingApp}
- * (child_process.spawn) with direct CLI arguments, bypassing protocol handlers
- * and their OS security prompts.
+ * Opens the host VS Code app from the Agents window via protocol handler.
+ * On desktop this action is replaced by the electron-browser override that
+ * uses {@link INativeHostService.launchSiblingApp} instead.
  */
 registerAction2(class OpenSessionWorktreeInVSCodeAction extends Action2 {
 	static readonly ID = 'chat.openSessionWorktreeInVSCode';
@@ -53,39 +51,58 @@ registerAction2(class OpenSessionWorktreeInVSCodeAction extends Action2 {
 		const telemetryService = accessor.get(ITelemetryService);
 		logSessionsInteraction(telemetryService, 'openInVSCode');
 
-		const nativeHostService = accessor.get(INativeHostService);
+		const openerService = accessor.get(IOpenerService);
 		const productService = accessor.get(IProductService);
 		const sessionsManagementService = accessor.get(ISessionsManagementService);
 		const sessionsProvidersService = accessor.get(ISessionsProvidersService);
 		const remoteAgentHostService = accessor.get(IRemoteAgentHostService);
 
+		const scheme = productService.quality === 'stable'
+			? 'vscode'
+			: productService.quality === 'exploration'
+				? 'vscode-exploration'
+				: productService.quality === 'insider'
+					? 'vscode-insiders'
+					: productService.urlProtocol;
+
+		const params = new URLSearchParams();
+		params.set('windowId', '_blank');
+
 		const activeSession = sessionsManagementService.activeSession.get();
-		const workspace = activeSession?.workspace.get();
+		if (!activeSession) {
+			await openerService.open(URI.from({ scheme, query: params.toString() }), { openExternal: true });
+			return;
+		}
+
+		const workspace = activeSession.workspace.get();
 		const repo = workspace?.repositories[0];
-		const rawFolderUri = activeSession?.sessionType === CopilotCLISessionType.id ? repo?.workingDirectory ?? repo?.uri : undefined;
-		const folderUri = rawFolderUri?.scheme === AGENT_HOST_SCHEME ? fromAgentHostUri(rawFolderUri) : rawFolderUri;
-		const remoteAuthority = activeSession
-			? resolveRemoteAuthority(activeSession.providerId, sessionsProvidersService, remoteAgentHostService)
-			: undefined;
+		const rawFolderUri = activeSession.sessionType === CopilotCLISessionType.id ? repo?.workingDirectory ?? repo?.uri : undefined;
 
-		const args: string[] = ['--new-window'];
-
-		if (folderUri) {
-			if (remoteAuthority) {
-				args.push('--folder-uri', URI.from({ scheme: Schemas.vscodeRemote, authority: remoteAuthority, path: folderUri.path }).toString());
-			} else {
-				args.push('--folder-uri', folderUri.toString());
-			}
+		if (!rawFolderUri) {
+			await openerService.open(URI.from({ scheme, query: params.toString() }), { openExternal: true });
+			return;
 		}
 
-		if (activeSession) {
-			const scheme = productService.parentPolicyConfig?.urlProtocol ?? productService.urlProtocol;
-			const params = new URLSearchParams();
-			params.set('windowId', '_blank');
-			params.set('session', activeSession.resource.toString());
-			args.push('--open-url', URI.from({ scheme, query: params.toString() }).toString());
-		}
+		const folderUri = rawFolderUri.scheme === AGENT_HOST_SCHEME ? fromAgentHostUri(rawFolderUri) : rawFolderUri;
+		const remoteAuthority = resolveRemoteAuthority(
+			activeSession.providerId, sessionsProvidersService, remoteAgentHostService);
 
-		await nativeHostService.launchSiblingApp(args);
+		params.set('session', activeSession.resource.toString());
+
+		if (remoteAuthority) {
+			await openerService.open(URI.from({
+				scheme,
+				authority: Schemas.vscodeRemote,
+				path: `/${remoteAuthority}${folderUri.path}`,
+				query: params.toString(),
+			}), { openExternal: true });
+		} else {
+			await openerService.open(URI.from({
+				scheme,
+				authority: Schemas.file,
+				path: folderUri.path,
+				query: params.toString(),
+			}), { openExternal: true });
+		}
 	}
 });
